+++ app/Catalog/Application/UseCases/CreateCategoryAttributeDefinition/CreateCategoryAttributeDefinitionHandler.php
@@ -26,6 +26,7 @@ readonly class CreateCategoryAttributeDefinitionHandler
             'placeholder'         => $input->placeholder,
             'help_text'           => $input->helpText,
             'default_value'       => $input->defaultValue,
+            'dependency_rules'    => $input->dependencyRules,
             'group_name'          => $input->groupName,
             'options'             => $input->options,
             'is_required'         => $input->isRequired,